updated projected routes and login page
